requirements.txt
tweepy==4.14.0
pytesseract==0.3.10
pillow==10.0.0
requests==2.31.0
beautifulsoup4==4.12.2
SQLAlchemy==2.0.17